feat: add task 1 (daftar semua pasien) and 2 (tambah pasien)

+++ sipas/src/main/java/apap/tugas/sipas/repository/PasienDb.java
@@ -3,9 +3,9 @@ package apap.tugas.sipas.repository;
 import apap.tugas.sipas.model.PasienModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.math.BigInteger;
+import java.util.Optional;
 
 @Repository
-public interface PasienDb extends JpaRepository<PasienModel, BigInteger> {
+public interface PasienDb extends JpaRepository<PasienModel, Long> {
+//    Optional<PasienModel> findByNIKPasien(Long NIKPasien);
 }
